# Path: optimize `join` for the case of joining one single part