Follow up to #8078

This optimizes `Path#join` (and `File.join`) for the most common case:

```crystal
File.join("prefix", "suffix")
```

When we just have two pieces to combine it becomes easier to compute the total bytesize needed for the string and we can use `String.new` instead of `String.build`, the latter involving an extra allocation for the `String::Builder` class (and there's also no need to write things though a pointer, check capacity, etc.)

This results in more code, but I think it's good if we optimize for this case because it's so common.

Benchmark code:

```crystal
require "benchmark"

base = "/Users/someone/foo/bar"
parts = ["baz.cr"]

Benchmark.ips do |x|
  x.report("join") do
    Path.new(base).join(parts)
  end
end
```

Results:

```
before: 10.90M ( 91.78ns) (± 2.80%)  144B/op
after:  19.14M ( 52.23ns) (±11.85%)  48.0B/op
```

That's twice as fast and with a third of the memory used.

Another benchmark (`Dir.glob` uses a lot of `File.join(x, y)`):

```crystal
require "benchmark"

Benchmark.ips do |x|
  x.report("Dir.glob") do
    Dir.glob("*/*.cr") do
    end
  end
end
```

Results:

```
before: 1.85k (541.76µs) (± 4.57%)  56.0kB/op
after:  1.89k (528.95µs) (±10.02%)  42.3kB/op
```

Not a lot of speed improvement because the `File.info?` call is the bottleneck (but that's solved by #8081), but look at the memory: 42.3kB/op vs. 56kB/op. Huge difference!